Fix as per Jonah's request: https://github.com/micro-nova/AmpliPi/pull/506#pullrequestreview-1605566389
